+++ packages/frontend/pages/settings/components/sidebar/SettingsSidebar.tsx
@@ -36,9 +36,11 @@ export const SettingsSidebar: React.FC<SettingsSidebarProps> = (props) => {
       </SettingsCategory>
 
       <SettingsCategory title="Account">
-        <SettingsLink href="/settings/info">Info</SettingsLink>
-        <SettingsLink href="/settings/password">Change password</SettingsLink>
-        <SettingsLink href="/settings/logout">Log out</SettingsLink>
+        <SettingsLink href="/settings/account/info">Info</SettingsLink>
+        <SettingsLink href="/settings/account/password">
+          Change password
+        </SettingsLink>
+        <SettingsLink href="/settings/account/logout">Log out</SettingsLink>
       </SettingsCategory>
 
       <SettingsCategory title="About">